requirements.txt
# Enterprise Strategy Playbook Platform - Python dependencies
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database / ORM
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0

# Data processing
pandas>=2.2.0
numpy>=1.26.3
pyarrow>=15.0.0

# LLM / AI
openai>=1.12.0
langchain>=0.1.6
langchain-openai>=0.0.5
tiktoken>=0.6.0
chromadb>=0.4.22

# Document generation
python-pptx>=0.6.23
fpdf2>=2.7.8
python-docx>=1.1.0
jinja2>=3.1.3

# HTTP / async
httpx>=0.27.0
aiohttp>=3.9.3

# Auth / security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Config / env
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1

# Utilities
rich>=13.7.0
structlog>=24.1.0
tenacity>=8.2.3

# HTML parsing
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

